Move unpack_shared_parameters_from_executor_and_invoke into its own header
